[mme] Encode APN AMBR to prevent 0x00 in octets 5 or 6
The spec for the variable precision apn-ambr encoding is complex and
somewhat ambiguous. In operating with a baicells nova 223 it appears
their decoder stops after encountering the value 0x00 in these bytes,
defined in TS24.301 9.9.4.2.1 as indicating "Use the value indicated
by the APN-AMBR for downlink in octet 3." This prevents the decoder
from processing octets 7 and 8 which may contain lower precision but
higher magnitude values.

This commit changes the encoding procedure to ensure that even if a
value could be encoded completely in octets 7 or 8 with no remainder
in octets 5 or 6, that the encoding prefers filling octets 5 or 6 with
at least some bits. Another alternative interpretation of the
spec *not* implemented in this commit could be to max out octets 5/6
before encoding any information in octets 7 or 8, but this would lead
to low precision around the common range of AMBRs from 256Mbps to a
few Gbps.